// src/app/shared/generated/model/mCPServer.ts
/**
 * onecx-ai bff
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ExecutionPolicy } from './executionPolicy';


export interface MCPServer { 
    modificationCount?: number;
    creationDate?: string;
    creationUser?: string;
    modificationDate?: string;
    modificationUser?: string;
    id?: string;
    name?: string;
    description?: string;
    url?: string;
    apiKey?: string;
    executionPolicy?: ExecutionPolicy;
}